pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "crawler"
version = "0.1.0"
description = "Asynchronous web crawler."
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.9",
    "aiofiles>=23.0",
    "aiosqlite>=0.19",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "pyyaml>=6.0",
]

[project.scripts]
crawler = "crawler.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
